requirements-render.txt
flask==3.0.0
gunicorn==21.2.0
python-telegram-bot==20.7
aiosqlite==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
solders==0.21.0
solana==0.30.2
spl-token==0.2.0
base58==2.1.1
